Cmor::All: Adapt to rails 6. Cmor::UserArea: Set minimum active_model_otp version to 2.1.1 as they added backup codes support that required changes to our code. Cmor::UserArea: Add one_time_password_options option as passthrough for active_model_otp options.

+++ cmor_cms/initialize_dummy.sh
@@ -7,10 +7,17 @@ rm -rf spec/dummy
 
 # Generate new dummy app
 DISABLE_MIGRATE=true rake dummy:app
-rm spec/dummy/.ruby-version
 
-# Satisfy prerequisites
+if [ ! -d "spec/dummy/config" ]; then exit 1; fi
+
+# Cleanup
+rm spec/dummy/.ruby-version
+rm spec/dummy/Gemfile
+
 cd spec/dummy
+
+# Use correct Gemfile
+sed -i "s|../Gemfile|../../../Gemfile|g" config/boot.rb
 
 # I18n configuration
 touch config/initializers/i18n.rb
@@ -22,6 +29,9 @@ touch config/initializers/route_translator.rb
 echo "RouteTranslator.config do |config|" >> config/initializers/route_translator.rb
 echo "  config.force_locale = true" >> config/initializers/route_translator.rb
 echo "end" >> config/initializers/route_translator.rb
+
+# Add webpacker
+sed -i '17i\  require "webpacker"' config/application.rb
 
 # Add ActiveStorage
 # rails active_storage:install
