+++ cmor_seo/initialize_dummy.sh
@@ -6,12 +6,14 @@ INSTALL_NAME=${GEM_NAME//cmor_/cmor\:}
 rm -rf spec/dummy
 
 # Generate new dummy app
-DISABLE_MIGRATE=true bundle exec rake dummy:app
+DISABLE_MIGRATE=true rake dummy:app
 
+if [ ! -d "spec/dummy/config" ]; then exit 1; fi
+
+# Cleanup
 rm spec/dummy/.ruby-version
 rm spec/dummy/Gemfile
 
-# Satisfy prerequisites
 cd spec/dummy
 
 # Use correct Gemfile
