+++ cmor_showcase_frontend/initialize_dummy.sh
@@ -9,14 +9,20 @@ rm -rf spec/dummy
 
 # Generate new dummy app
 DISABLE_MIGRATE=true rake dummy:app
-rm spec/dummy/.ruby-version
 
-# Satisfy prerequisites
+if [ ! -d "spec/dummy/config" ]; then exit 1; fi
+
+# Cleanup
+rm spec/dummy/.ruby-version
+rm spec/dummy/Gemfile
+
 cd spec/dummy
 
 # Use correct Gemfile
-rm Gemfile
 sed -i "s|../Gemfile|../../../Gemfile|g" config/boot.rb
+
+# Use webpacker
+sed -i '17irequire "webpacker"' config/application.rb
 
 # Add ActiveStorage
 rails active_storage:install
