+++ cmor_tags/initialize_dummy.sh
@@ -7,18 +7,25 @@ rm -rf spec/dummy
 
 # Generate new dummy app
 DISABLE_MIGRATE=true rake dummy:app
-rm spec/dummy/.ruby-version
 
-# Satisfy prerequisites
+if [ ! -d "spec/dummy/config" ]; then exit 1; fi
+
+# Cleanup
+rm spec/dummy/.ruby-version
+rm spec/dummy/Gemfile
+
 cd spec/dummy
 
+# Use correct Gemfile
+sed -i "s|../Gemfile|../../../Gemfile|g" config/boot.rb
+
 # Responders for rao-service_controller
-sed -i '17i\  require "responders"' config/application.rb
+sed -i '17i\require "responders"' config/application.rb
 
 # Require needed stuff dummy app
-sed -i '17i\  require "puma"' config/application.rb
-sed -i '17i\  require "rspec-rails"' config/application.rb
-sed -i '17i\  require "factory_bot_rails"' config/application.rb
+sed -i '17i\require "webpacker"' config/application.rb
+sed -i '17i\require "rspec-rails"' config/application.rb
+sed -i '17i\require "factory_bot_rails"' config/application.rb
 
 # I18n configuration
 touch config/initializers/i18n.rb
